backend/requirements.txt
Flask
Flask-Cors
Flask-SocketIO
pandas
joblib
scikit-learn
shap
python-whois
dnspython
pyOpenSSL
requests
python-levenshtein
fpdf
pyzbar
opencv-python-headless
beautifulsoup4
eventlet
tldextract

tensorflow
scikit-learn
